+++ packages/core/src/components/Route.tsx
@@ -1,28 +1,44 @@
 import React from 'react';
 import { useRouterContext } from '../context';
 import { getPathParams } from '../helpers';
-import { NotFound } from '.';
 
 interface RouteProps {
   path: string;
   component: React.ComponentType<IComponentProps>;
-  notFoundPage?: React.ComponentType<any>;
+  basePath?: string;
+  children?: React.ReactNode;
 }
 
 interface IComponentProps {
   pathParams: object;
   queryParams: object;
+  children?: React.ReactElement<IComponentProps>[] | null;
+  basePath?: string;
 }
 
-export const Route: React.FC<RouteProps> = ({ path, component: Component, notFoundPage: NotFoundPage }) => {
+export const Route:
+React.FC<RouteProps> = ({
+  path, component: Component, basePath, children,
+}) => {
   const { path: currentPath, queryParams } = useRouterContext();
-  const pathParams = getPathParams(path, currentPath);
+  const fullPath = basePath ? `${basePath}${path}` : path;
+  const pathParams = getPathParams(fullPath, currentPath);
 
-  const isMatch = path === "*" ? true : currentPath.startsWith(path);
+  const isMatch = fullPath === '*' ? true : currentPath.startsWith(fullPath);
 
   if (!isMatch) return null;
 
-  if (path === "*") return NotFoundPage ? <NotFoundPage /> : <NotFound />;
-
-  return <Component pathParams={pathParams} queryParams={queryParams} />;
+  return (
+    <Component pathParams={pathParams} queryParams={queryParams}>
+      {React.Children.map(
+        children as React.ReactElement<IComponentProps>[] | null,
+        (child: React.ReactElement<IComponentProps> | null | undefined) => {
+          if (React.isValidElement(child)) {
+            return React.cloneElement(child, { basePath: fullPath });
+          }
+          return child;
+        },
+      )}
+    </Component>
+  );
 };