+++ packages/core/src/components/index.ts
@@ -2,4 +2,6 @@ export * from './Link';
 export * from './Route';
 export * from './Router';
 export * from './Redirect';
+export * from './Switch';
+
 export * from './pages';