+++ packages/core/src/components/pages/NotFound.tsx
@@ -1,8 +1,6 @@
-export const NotFound = () => {
-  return (
-    <div>
-      <h1>404</h1>
-      <p>Page not found</p>
-    </div>
-  )
-}
+export const NotFound = () => (
+  <div>
+    <h1>404</h1>
+    <p>Page not found</p>
+  </div>
+);