+++ packages/core/src/context/RouterContext.tsx
@@ -32,7 +32,10 @@ React.FC<RouterContextProviderProps> = ({ children }) => {
   window.onpopstate = () => setLocation(window.location.pathname);
 
   return (
-    <RouterContext.Provider value={{ path: location, setLocation, queryParams, navigationData, setNavigationData }}>
+    <RouterContext.Provider value={{
+      path: location, setLocation, queryParams, navigationData, setNavigationData,
+    }}
+    >
       {children}
     </RouterContext.Provider>
   );
@@ -41,28 +44,25 @@ React.FC<RouterContextProviderProps> = ({ children }) => {
 export const useRouterContext = () => {
   const context = React.useContext(RouterContext);
 
-  const { path, setLocation, queryParams, navigationData, setNavigationData } = context ?? {};
+  const {
+    path, setLocation, queryParams, navigationData, setNavigationData,
+  } = context ?? {};
 
   const navigate = (path: string, data?: object) => {
     window.history.pushState(null, '', path);
     setLocation(path);
     setNavigationData(data ?? {});
-    return;
   };
 
   const goBack = () => {
     window.history.back();
-    return;
-  }
+  };
 
   const goForward = () => {
     window.history.forward();
-    return;
-  }
+  };
 
-  const getNavigationData = () => {
-    return navigationData;
-  }
+  const getNavigationData = () => navigationData;
 
   return {
     path,