+++ packages/core/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @pedrovs3/react-router
 
+## 1.2.6
+
+### Patch Changes
+
+- fix validate of path segments in switch component
+
+## 1.2.5
+
+### Patch Changes
+
+- corrections on switch component
+
+## 1.2.4
+
+### Patch Changes
+
+- corrections to nested routes with switch component
+
+## 1.2.3
+
+### Patch Changes
+
+- corrections in nested routes (switch component)
+
+## 1.2.2
+
+### Patch Changes
+
+- corrections on switch
+
+## 1.2.1
+
+### Patch Changes
+
+- corrections on switch component
+
+## 1.2.0
+
+### Minor Changes
+
+- Switch component
+
 ## 1.1.0
 
 ### Minor Changes